This "works", but it's wrong
Presently it's using the index to pull files from the repo,
and thereby getting anything that ever had a reference.

I'd much rather show the current state of master,
so I'll be changing it to do that.

Also, it's looking up the files in the file system,
which is bad because this can't be used on bare repos,
 and since I'm using the index it's looking for files that don't exist.